requirements.txt
requests
openai
tiktoken
python-dotenv